Nothing works...

+++ src/main/java/com/dk_power/power_plant_java/sevice/angular/NgValueService.java
@@ -5,6 +5,8 @@ import com.dk_power.power_plant_java.dto.categories.ValueDto;
 import com.dk_power.power_plant_java.entities.categories.Category;
 import com.dk_power.power_plant_java.entities.categories.Value;
 import com.dk_power.power_plant_java.entities.equipment.Equipment;
+import com.dk_power.power_plant_java.entities.files.FileObject;
+import com.dk_power.power_plant_java.entities.loto.LotoPoint;
 import com.dk_power.power_plant_java.repository.categories.CategoryRepo;
 import com.dk_power.power_plant_java.repository.categories.ValueRepo;
 import com.dk_power.power_plant_java.sevice.angular.file.NgFileService;
@@ -124,7 +126,12 @@ public class NgValueService {
     public void deleteValue(Long id) {
         Value value = getValueById(id).orElseThrow(() -> new RuntimeException("Value not found"));
         List<Equipment> associatedEq = equipmentService.findByValue(value);
-        valueRepo.deleteById(id);
+        List<FileObject> associatedFiles = fileService.findByValue(value);
+        List<LotoPoint> associatedLotoPoints = lotoPointService.findByValue(value);
+        if(associatedLotoPoints.isEmpty() && associatedFiles.isEmpty() && associatedEq.isEmpty()){
+            valueRepo.deleteById(id);
+        }
+
     }
 
     // Move value to different category
